chore: added process env to npm scripts

+++ gulp-tasks/build-html.js
@@ -9,8 +9,12 @@ const data = require('./../src/data')
 const addBundleCssJs = (data) => {
   data['bundleCss'] = fs.readFileSync('./build/css/bundle.css')
   data['bundleJs'] = fs.readFileSync('./build/js/bundle.js')
-  fs.rmdirSync('./build/css', { recursive: true })
-  fs.rmdirSync('./build/js', { recursive: true })
+
+  if (process.env.NODE_ENV == 'prod') {
+    fs.rmdirSync('./build/css', { recursive: true })
+    fs.rmdirSync('./build/js', { recursive: true })
+  }
+
   return data
 }
